microservices/device/create_device/requirements.txt
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5
pydantic==1.6.1
aws-lambda-powertools==1.8.0
pycryptodome